Fix structure changes of files array
The restructuring of the `$_FILES` will fail if my uploaded file location has a various depth (e.h. `form[fields]['images']['header']`). The current implementation would create a result like this. It picks the first key of the path and uses it as field name.

```
array (
  'fields' => array(
    'name' => array(
      'images' => array(
        // ...
      )
    ),
    'tmp_name' => '...',
    'size' => '...',
  )
)
```
Instead I expect to get an array like this: 
```
array (
  'fields' => array(
    'images' => array(
      'header' => array(
        'name' => 'filename.txt',
        'tmp_name' => '...',
        'size' => '...',
      )
    )
    'name' => array(
      'images' => array(
        // ...
      )
    ),
    'tmp_name' => '...',
    'size' => '...',
  )
)
```

This is what the PR will do.